Reorganize controllers to follow Rails guidelines.
