+++ app/controllers/champions_controller.rb
@@ -1,0 +1,5 @@
+class ChampionsController < ApplicationController
+  def by_win_rate
+
+  end
+end
