+++ app/views/champions/by_win_rate.html.haml
